CR: Fix units & comments
