+++ src/Polly.Shared/Retry/DecorrelatedJitterBackoff.cs
@@ -14,8 +14,10 @@ namespace Polly.Retry
         private readonly ConcurrentRandom _random;
 
         /// <summary>
-        /// Whether the first retry will be immediate or not.
+        /// Specifies whether the first retry will be immediate or not.
         /// </summary>
+        /// <remarks>When true, a first retry is made after zero delay, after which the configured
+        /// backoff strategy is followed. The overall number of retries is not increased.</remarks>
         public bool FastFirst { get; }
 
         /// <summary>
