+++ src/Polly.Shared/Retry/ISleepDurationSeriesStrategy.cs
@@ -9,8 +9,10 @@ namespace Polly.Retry
     public interface ISleepDurationSeriesStrategy
     {
         /// <summary>
-        /// Whether the first retry will be immediate or not.
+        /// Specifies whether the first retry will be immediate or not.
         /// </summary>
+        /// <remarks>When true, a first retry is made after zero delay, after which the configured
+        /// backoff strategy is followed. The overall number of retries is not increased.</remarks>
         bool FastFirst { get; }
 
         /// <summary>
