+++ src/Polly.SharedSpecs/Retry/DecorrelatedJitterBackoffSpecs.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Xunit;
 
-namespace Polly.Specs.Duration
+namespace Polly.Specs.Retry
 {
     public static class DecorrelatedJitterBackoffSpecs
     {
